requirements.txt
Django==1.11.2
django-debug-toolbar==1.8
django-extensions==1.7.9
djangorestframework==3.6.3
psycopg2==2.6.2
PyJWT==1.5.2
coverage==4.4.1
Sphinx==1.6.2
elizabeth==0.3.11
python-magic==0.4.13

gunicorn==19.7.1

requests